Create a basic pickup system
Hey you need a basic pick up system for your 3d project in Godot but don't know where to start? Don't fret I can show you the basics! If you have any issues please join the discord link is provided at the bottom of the code and ask away!

This was created for Godot 4.2.2 and may not work in later or previous versions!